Change register to register for all shifts in stack
